Changes to build with cmake/visual studio. Still have link error with readline5.dll
git-svn-id: https://simtk.org/svn/sequoia/branches/rel1_0@126 3da64b0a-0202-0410-8c91-c753549fdeb7

+++ src/consens.hxx
@@ -13,6 +13,6 @@
 
 Sequence consensus(const Alignment & a, 
 		   const Comparison_matrix & cm, 
-		   float cutoff);
+		   double cutoff);
 
 #endif
